Add relation on model
