+++ app/Models/Arsip/Arsip.php
@@ -2,10 +2,22 @@
 
 namespace App\Models\Arsip;
 
+use App\Models\User\User;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Relations\BelongsTo;
 
 class Arsip extends Model
 {
     use HasFactory;
+
+    /**
+     * Get the user that owns the Arsip
+     *
+     * @return \Illuminate\Database\Eloquent\Relations\BelongsTo
+     */
+    public function user(): BelongsTo
+    {
+        return $this->belongsTo(User::class);
+    }
 }
