+++ app/Models/Surat/Surat.php
@@ -4,8 +4,19 @@ namespace App\Models\Surat;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Relations\BelongsTo;
 
 class Surat extends Model
 {
     use HasFactory;
+
+    /**
+     * Get the user that owns the Surat
+     *
+     * @return \Illuminate\Database\Eloquent\Relations\BelongsTo
+     */
+    public function user(): BelongsTo
+    {
+        return $this->belongsTo(User::class);
+    }
 }
